third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/style-change-in-container.html

<!DOCTYPE html>
<title>Container Queries: Recalculate style inside a container</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #container { contain: size layout }
  @container (min-width: 1px) {
    #content { color: green; }
  }
</style>
<div id=container>
  <div id=content>
    Test
  </div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(content).color, 'rgb(0, 128, 0)');

    // Dirty style of an element inside the container:
    content.style.backgroundColor = 'lightgray';

    // The container query should still evaluate correctly:
    assert_equals(getComputedStyle(content).color, 'rgb(0, 128, 0)');
  }, 'Hello');
</script>
